Horarios Medicos completado

+++ routes/web.php
@@ -15,24 +15,36 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', function () {
-    return view('auth.login');
-});
+     return view('auth.login');
+ });
+// Route::get('/', function () {
+//     return view('welcome');
+// });
 
 Auth::routes();
-
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-//rutas especialidades
-Route::get('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'index']);
-Route::get('/especialidades/create', [App\Http\Controllers\SpecialtyController::class, 'create']);
-Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\SpecialtyController::class, 'edit']);
-Route::post('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'sendData']);
 
-Route::put('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'update']);
+Route::middleware(['auth', 'admin'])->group(function () {
+    
+    //rutas especialidades
+    Route::get('/especialidades', [App\Http\Controllers\Admin\SpecialtyController::class, 'index']);
+    Route::get('/especialidades/create', [App\Http\Controllers\Admin\SpecialtyController::class, 'create']);
+    Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\Admin\SpecialtyController::class, 'edit']);
+    Route::post('/especialidades', [App\Http\Controllers\Admin\SpecialtyController::class, 'sendData']);
 
-Route::delete('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'destroy']);
+    Route::put('/especialidades/{specialty}', [App\Http\Controllers\Admin\SpecialtyController::class, 'update']);
 
-//rutas medicos
-Route::resource('medicos', 'App\Http\Controllers\DoctorController');
+    Route::delete('/especialidades/{specialty}', [App\Http\Controllers\Admin\SpecialtyController::class, 'destroy']);
 
-//rutas pacientes
-Route::resource('pacientes', 'App\Http\Controllers\PatientController');
+    //rutas medicos
+    Route::resource('medicos', 'App\Http\Controllers\Admin\DoctorController');
+
+    //rutas pacientes
+    Route::resource('pacientes', 'App\Http\Controllers\Admin\PatientController');
+});
+
+Route::middleware(['auth', 'doctor'])->group(function () {
+    Route::get('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'edit']);
+    Route::post('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'store']);
+
+});
